moving rexdb to src directory
